requirements.txt
paddlepaddle
pandas~=2.2.3
numpy~=2.0.2
pyyaml~=6.0.2

scikit-learn~=1.5.2